move the controller handle into a separate module for clarity
